requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Optional but recommended
scipy>=1.7.0
scikit-image>=0.19.0
pyyaml>=6.0
tqdm>=4.62.0

# For training
tensorboard>=2.10.0

# For advanced features (optional)
# rawpy>=0.18.0  # RAW image processing
# pywavelets>=1.3.0  # Wavelet transforms

# Development
# pytest>=7.0.0
# black>=22.0.0
# isort>=5.10.0
